+++ plugins/preference/src/android/support/v4/app/SherlockPreferenceActivity.java
@@ -19,42 +19,76 @@ package android.support.v4.app;
 
 import java.io.FileDescriptor;
 import java.io.PrintWriter;
-import android.app.Activity;
-import android.content.Intent;
-import android.content.res.Resources.Theme;
-import android.content.res.TypedArray;
-import android.os.Build;
-import android.os.Bundle;
-import android.os.Handler;
-import android.preference.PreferenceActivity;
-import android.support.v4.view.ActionMode;
-import android.support.v4.view.Menu;
-import android.support.v4.view.MenuInflater;
-import android.support.v4.view.MenuItem;
-import android.support.v4.view.Window;
-import android.util.Log;
-import android.view.View;
-import android.view.ViewGroup;
-import android.view.ViewGroup.LayoutParams;
-import android.widget.ListView;
+import java.util.ArrayList;
+import java.util.HashMap;
 import com.actionbarsherlock.R;
-import com.actionbarsherlock.internal.app.ActionBarImpl;
 import com.actionbarsherlock.internal.app.ActionBarWrapper;
+import com.actionbarsherlock.internal.app.ActionBarImpl;
 import com.actionbarsherlock.internal.view.menu.MenuBuilder;
 import com.actionbarsherlock.internal.view.menu.MenuInflaterWrapper;
 import com.actionbarsherlock.internal.view.menu.MenuItemImpl;
 import com.actionbarsherlock.internal.view.menu.MenuItemWrapper;
 import com.actionbarsherlock.internal.view.menu.MenuWrapper;
 
+import android.R.menu;
+import android.app.Activity;
+import android.content.Context;
+import android.content.Intent;
+import android.content.res.Configuration;
+import android.content.res.TypedArray;
+import android.content.res.Resources.Theme;
+import android.os.Build;
+import android.os.Bundle;
+import android.os.Handler;
+import android.preference.PreferenceActivity;
+import android.os.Message;
+import android.os.Parcelable;
+import android.support.v4.view.ActionMode;
+import android.support.v4.view.Menu;
+import android.support.v4.view.MenuInflater;
+import android.support.v4.view.MenuItem;
+import android.support.v4.view.Window;
+import android.util.AttributeSet;
+import android.util.Log;
+import android.view.KeyEvent;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.view.ViewGroup.LayoutParams;
+import android.widget.FrameLayout;
+import android.widget.ListView;
+
 /**
- * Base class for activities that want to use the support-based ActionBar and
- * Preference APIs.
+ * Base class for activities that want to use the support-based ActionBar,
+ * Fragment, Loader, and Preference APIs.
+ *
+ * <p>Known limitations:</p>
+ * <ul>
+ * <li> <p>When using the &lt;fragment> tag, this implementation can not
+ * use the parent view's ID as the new fragment's ID.  You must explicitly
+ * specify an ID (or tag) in the &lt;fragment>.</p>
+ * <li> <p>Prior to Honeycomb (3.0), an activity's state was saved before pausing.
+ * Fragments are a significant amount of new state, and dynamic enough that one
+ * often wants them to change between pausing and stopping.  These classes
+ * throw an exception if you try to change the fragment state after it has been
+ * saved, to avoid accidental loss of UI state.  However this is too restrictive
+ * prior to Honeycomb, where the state is saved before pausing.  To address this,
+ * when running on platforms prior to Honeycomb an exception will not be thrown
+ * if you change fragments between the state save and the activity being stopped.
+ * This means that is some cases if the activity is restored from its last saved
+ * state, this may be a snapshot slightly before what the user last saw.</p>
+ * </ul>
  */
-public class SherlockPreferenceActivity extends PreferenceActivity implements SupportActivity {
+public abstract class SherlockPreferenceActivity extends PreferenceActivity implements SupportActivity {
     private static final String TAG = "SherlockPreferenceActivity";
     private static final boolean DEBUG = false;
 
+    private static final String FRAGMENTS_TAG = "android:support:fragments";
+
     static final boolean IS_HONEYCOMB = Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB;
+
+    static final int MSG_REALLY_STOPPED = 1;
+    static final int MSG_RESUME_PENDING = 2;
 
     private static final int WINDOW_FLAG_ACTION_BAR = 1 << Window.FEATURE_ACTION_BAR;
     private static final int WINDOW_FLAG_ACTION_BAR_ITEM_TEXT = 1 << Window.FEATURE_ACTION_BAR_ITEM_TEXT;
@@ -65,22 +99,22 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
     final SupportActivity.InternalCallbacks mInternalCallbacks = new SupportActivity.InternalCallbacks() {
         @Override
         void invalidateSupportFragmentIndex(int index) {
-            //No op
+            SherlockPreferenceActivity.this.invalidateSupportFragmentIndex(index);
         }
 
         @Override
         LoaderManagerImpl getLoaderManager(int index, boolean started, boolean create) {
-            return null;
+            return SherlockPreferenceActivity.this.getLoaderManager(index, started, create);
         }
 
         @Override
         Handler getHandler() {
-            return null;
+            return mHandler;
         }
 
         @Override
         FragmentManagerImpl getFragments() {
-            return null;
+            return mFragments;
         }
 
         @Override
@@ -90,9 +124,29 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
 
         @Override
         boolean getRetaining() {
-            return false;
+            return mRetaining;
         }
     };
+
+    final Handler mHandler = new Handler() {
+        @Override
+        public void handleMessage(Message msg) {
+            switch (msg.what) {
+                case MSG_REALLY_STOPPED:
+                    if (mStopped) {
+                        doReallyStop(false);
+                    }
+                    break;
+                case MSG_RESUME_PENDING:
+                    mFragments.dispatchResume();
+                    mFragments.execPendingActions();
+                    break;
+                default:
+                    super.handleMessage(msg);
+            }
+        }
+    };
+    final FragmentManagerImpl mFragments = new FragmentManagerImpl();
 
     ActionBar mActionBar;
     boolean mIsActionBarImplAttached;
@@ -106,8 +160,37 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
         }
     };
 
+    boolean mCreated;
+    boolean mResumed;
+    boolean mStopped;
+    boolean mReallyStopped;
+    boolean mRetaining;
+
     boolean mOptionsMenuInvalidated;
     boolean mOptionsMenuCreateResult;
+
+    boolean mCheckedForLoaderManager;
+    boolean mLoadersStarted;
+    HCSparseArray<LoaderManagerImpl> mAllLoaderManagers;
+    LoaderManagerImpl mLoaderManager;
+
+    static final class NonConfigurationInstances {
+        Object activity;
+        Object custom;
+        HashMap<String, Object> children;
+        ArrayList<Fragment> fragments;
+        HCSparseArray<LoaderManagerImpl> loaders;
+    }
+
+    static class FragmentTag {
+        public static final int[] Fragment = {
+            0x01010003, 0x010100d0, 0x010100d1
+        };
+        public static final int Fragment_id = 1;
+        public static final int Fragment_name = 0;
+        public static final int Fragment_tag = 2;
+    }
+
 
 
     public SherlockPreferenceActivity() {
@@ -131,25 +214,26 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
     public Activity asActivity() {
         return this;
     }
+    
+    boolean mEnsureInflating = false;
 
     protected void ensureSupportActionBarAttached() {
         if (IS_HONEYCOMB || mIsActionBarImplAttached) {
             return;
         }
         if (!isChild()) {
-            final ListView contentView = new ListView(this);
-            contentView.setLayoutParams(new LayoutParams(LayoutParams.FILL_PARENT, LayoutParams.FILL_PARENT));
-            contentView.setId(android.R.id.list);
-
+        	//Get the list view that the parent attached
+            final ListView contentView = (ListView)getWindow().getDecorView().findViewById(android.R.id.list);
+            ((ViewGroup)contentView.getParent()).removeView(contentView);
+            
+            //Disable the callback in PrefrenceActivity
+            mEnsureInflating = true;
+            
             if ((mWindowFlags & WINDOW_FLAG_ACTION_BAR) == WINDOW_FLAG_ACTION_BAR) {
                 if ((mWindowFlags & WINDOW_FLAG_ACTION_BAR_OVERLAY) == WINDOW_FLAG_ACTION_BAR_OVERLAY) {
-                    View view = getLayoutInflater().inflate(R.layout.abs__screen_action_bar_overlay, null);
-                    ((ViewGroup)view.findViewById(R.id.abs__content)).addView(contentView);
-                    super.setContentView(view);
+                    super.setContentView(R.layout.abs__screen_action_bar_overlay);
                 } else {
-                    View view = getLayoutInflater().inflate(R.layout.abs__screen_action_bar, null);
-                    ((ViewGroup)view.findViewById(R.id.abs__content)).addView(contentView);
-                    super.setContentView(view);
+                    super.setContentView(R.layout.abs__screen_action_bar);
                 }
 
                 mActionBar = new ActionBarImpl(this);
@@ -167,21 +251,33 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
                 if ((mWindowFlags & WINDOW_FLAG_INDETERMINANTE_PROGRESS) == WINDOW_FLAG_INDETERMINANTE_PROGRESS) {
                     super.requestWindowFeature((int)Window.FEATURE_INDETERMINATE_PROGRESS);
                 }
-                View view = getLayoutInflater().inflate(R.layout.abs__screen_simple, null);
-                ((ViewGroup)view.findViewById(R.id.abs__content)).addView(contentView);
-                super.setContentView(view);
+                super.setContentView(R.layout.abs__screen_simple);
             }
+
+            //Attach original list view to the new layout
+            ((ViewGroup)findViewById(R.id.abs__content)).addView(contentView);
+            
+            //Re-enable the callback in PreferenceActivity
+            mEnsureInflating = false;
+            super.onContentChanged();
         }
 
         invalidateOptionsMenu();
         mIsActionBarImplAttached = true;
     }
 
+    @Override
+	public void onContentChanged() {
+		if (!mEnsureInflating) {
+			super.onContentChanged();
+		}
+	}
+
     // ------------------------------------------------------------------------
     // HOOKS INTO ACTIVITY
     // ------------------------------------------------------------------------
 
-    /**
+	/**
      * Enable extended window features.
      *
      * @param featureId The desired feature as defined in
@@ -220,6 +316,7 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
         return new MenuInflater(this, super.getMenuInflater());
     }
 
+    @Override
     public void setTitle(CharSequence title) {
         if (IS_HONEYCOMB || (getSupportActionBar() == null)) {
             super.setTitle(title);
@@ -237,6 +334,32 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
         }
     }
 
+    /**
+     * Dispatch incoming result to the correct fragment.
+     */
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        int index = requestCode>>16;
+        if (index != 0) {
+            index--;
+            if (mFragments.mActive == null || index < 0 || index >= mFragments.mActive.size()) {
+                Log.w(TAG, "Activity result fragment index out of range: 0x"
+                        + Integer.toHexString(requestCode));
+                return;
+            }
+            Fragment frag = mFragments.mActive.get(index);
+            if (frag == null) {
+                Log.w(TAG, "Activity result no fragment exists for index: 0x"
+                        + Integer.toHexString(requestCode));
+            } else {
+                frag.onActivityResult(requestCode&0xffff, resultCode, data);
+            }
+            return;
+        }
+
+        super.onActivityResult(requestCode, resultCode, data);
+    }
+
     @Override
     protected void onApplyThemeResource(Theme theme, int resid, boolean first) {
         TypedArray attrs = theme.obtainStyledAttributes(resid, R.styleable.SherlockTheme);
@@ -251,10 +374,50 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
         super.onApplyThemeResource(theme, resid, first);
     }
 
+    /**
+     * Take care of popping the fragment back stack or finishing the activity
+     * as appropriate.
+     */
+    @Override
+    public void onBackPressed() {
+        if (!mFragments.popBackStackImmediate()) {
+            finish();
+        }
+    }
+
+    /**
+     * Dispatch configuration change to all fragments.
+     */
+    @Override
+    public void onConfigurationChanged(Configuration newConfig) {
+        super.onConfigurationChanged(newConfig);
+        mFragments.dispatchConfigurationChanged(newConfig);
+    }
+
+    /**
+     * Perform initialization of all fragments and loaders.
+     */
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+        mFragments.attachActivity(this);
+        // Old versions of the platform didn't do this!
+        if (getLayoutInflater().getFactory() == null) {
+            getLayoutInflater().setFactory(this);
+        }
+
         super.onCreate(savedInstanceState);
+
+        NonConfigurationInstances nc = (NonConfigurationInstances)
+                getLastNonConfigurationInstance();
+        if (nc != null) {
+            mAllLoaderManagers = nc.loaders;
+        }
+        if (savedInstanceState != null) {
+            Parcelable p = savedInstanceState.getParcelable(FRAGMENTS_TAG);
+            mFragments.restoreAllState(p, nc != null ? nc.fragments : null);
+        }
         ensureSupportActionBarAttached();
+        mFragments.dispatchCreate();
     }
 
     /**
@@ -295,11 +458,95 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
         if (IS_HONEYCOMB) {
             if (DEBUG) Log.d(TAG, "onCreateOptionsMenu(android.view.Menu): Calling support method with wrapped native menu.");
             MenuWrapper wrapped = new MenuWrapper(menu);
-            result = onCreateOptionsMenu(wrapped);
+            result  = onCreateOptionsMenu(wrapped);
+            result |= mFragments.dispatchCreateOptionsMenu(wrapped, getMenuInflater());
         }
 
         if (DEBUG) Log.d(TAG, "onCreateOptionsMenu(android.view.Menu): Returning " + result);
         return result;
+    }
+
+    /**
+     * Add support for inflating the &lt;fragment> tag.
+     */
+    @Override
+    public View onCreateView(String name, Context context, AttributeSet attrs) {
+        if (!"fragment".equals(name)) {
+            return super.onCreateView(name, context, attrs);
+        }
+
+        String fname = attrs.getAttributeValue(null, "class");
+        TypedArray a =  context.obtainStyledAttributes(attrs, FragmentTag.Fragment);
+        if (fname == null) {
+            fname = a.getString(FragmentTag.Fragment_name);
+        }
+        int id = a.getResourceId(FragmentTag.Fragment_id, View.NO_ID);
+        String tag = a.getString(FragmentTag.Fragment_tag);
+        a.recycle();
+
+        View parent = null; // NOTE: no way to get parent pre-Honeycomb.
+        int containerId = parent != null ? parent.getId() : 0;
+        if (containerId == View.NO_ID && id == View.NO_ID && tag == null) {
+            throw new IllegalArgumentException(attrs.getPositionDescription()
+                    + ": Must specify unique android:id, android:tag, or have a parent with an id for " + fname);
+        }
+
+        // If we restored from a previous state, we may already have
+        // instantiated this fragment from the state and should use
+        // that instance instead of making a new one.
+        Fragment fragment = id != View.NO_ID ? mFragments.findFragmentById(id) : null;
+        if (fragment == null && tag != null) {
+            fragment = mFragments.findFragmentByTag(tag);
+        }
+        if (fragment == null && containerId != View.NO_ID) {
+            fragment = mFragments.findFragmentById(containerId);
+        }
+
+        if (FragmentManagerImpl.DEBUG) Log.v(TAG, "onCreateView: id=0x"
+                + Integer.toHexString(id) + " fname=" + fname
+                + " existing=" + fragment);
+        if (fragment == null) {
+            fragment = Fragment.instantiate(this, fname);
+            fragment.mFromLayout = true;
+            fragment.mFragmentId = id != 0 ? id : containerId;
+            fragment.mContainerId = containerId;
+            fragment.mTag = tag;
+            fragment.mInLayout = true;
+            fragment.mFragmentManager = mFragments;
+            fragment.onInflate(this, attrs, fragment.mSavedFragmentState);
+            mFragments.addFragment(fragment, true);
+
+        } else if (fragment.mInLayout) {
+            // A fragment already exists and it is not one we restored from
+            // previous state.
+            throw new IllegalArgumentException(attrs.getPositionDescription()
+                    + ": Duplicate id 0x" + Integer.toHexString(id)
+                    + ", tag " + tag + ", or parent id 0x" + Integer.toHexString(containerId)
+                    + " with another fragment for " + fname);
+        } else {
+            // This fragment was retained from a previous instance; get it
+            // going now.
+            fragment.mInLayout = true;
+            // If this fragment is newly instantiated (either right now, or
+            // from last saved state), then give it the attributes to
+            // initialize itself.
+            if (!fragment.mRetaining) {
+                fragment.onInflate(this, attrs, fragment.mSavedFragmentState);
+            }
+            mFragments.moveToState(fragment);
+        }
+
+        if (fragment.mView == null) {
+            throw new IllegalStateException("Fragment " + fname
+                    + " did not create a view.");
+        }
+        if (id != 0) {
+            fragment.mView.setId(id);
+        }
+        if (fragment.mView.getTag() == null) {
+            fragment.mView.setTag(tag);
+        }
+        return fragment.mView;
     }
 
     @Override
@@ -311,10 +558,13 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
         } else {
             mSupportMenu.clear();
 
-            mOptionsMenuCreateResult = onCreateOptionsMenu(mSupportMenu);
+            mOptionsMenuCreateResult  = onCreateOptionsMenu(mSupportMenu);
+            mOptionsMenuCreateResult |= mFragments.dispatchCreateOptionsMenu(mSupportMenu, getMenuInflater());
 
             if (getSupportActionBar() != null) {
-                onPrepareOptionsMenu(mSupportMenu);
+                if (onPrepareOptionsMenu(mSupportMenu)) {
+                    mFragments.dispatchPrepareOptionsMenu(mSupportMenu);
+                }
 
                 //Since we now know we are using a custom action bar, perform the
                 //inflation callback to allow it to display any items it wants.
@@ -333,9 +583,84 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
         }
     }
 
+    /**
+     * Destroy all fragments and loaders.
+     */
+    @Override
+    protected void onDestroy() {
+        super.onDestroy();
+
+        doReallyStop(false);
+
+        mFragments.dispatchDestroy();
+        if (mLoaderManager != null) {
+            mLoaderManager.doDestroy();
+        }
+    }
+
+    /**
+     * Take care of calling onBackPressed() for pre-Eclair platforms.
+     */
+    @Override
+    public boolean onKeyDown(int keyCode, KeyEvent event) {
+        if (android.os.Build.VERSION.SDK_INT < 5 /* ECLAIR */
+                && keyCode == KeyEvent.KEYCODE_BACK
+                && event.getRepeatCount() == 0) {
+            // Take care of calling this method on earlier versions of
+            // the platform where it doesn't exist.
+            onBackPressed();
+            return true;
+        }
+
+        return super.onKeyDown(keyCode, event);
+    }
+
+    /**
+     * Dispatch onLowMemory() to all fragments.
+     */
+    @Override
+    public void onLowMemory() {
+        super.onLowMemory();
+        mFragments.dispatchLowMemory();
+    }
+
+    /**
+     * Dispatch context and options menu to fragments.
+     */
+    @Override
+    public final boolean onMenuItemSelected(int featureId, android.view.MenuItem item) {
+        if (super.onMenuItemSelected(featureId, item)) {
+            return true;
+        }
+
+        switch (featureId) {
+            case Window.FEATURE_OPTIONS_PANEL:
+                return mFragments.dispatchOptionsItemSelected(new MenuItemWrapper(item));
+
+            case Window.FEATURE_CONTEXT_MENU:
+                return mFragments.dispatchContextItemSelected(new MenuItemWrapper(item));
+
+            default:
+                return false;
+        }
+    }
+
     @Override
     public boolean onMenuItemSelected(int featureId, MenuItem item) {
-        return super.onMenuItemSelected(featureId, item);
+        if (onOptionsItemSelected(item)) {
+            return true;
+        }
+
+        switch (featureId) {
+            case Window.FEATURE_OPTIONS_PANEL:
+                return mFragments.dispatchOptionsItemSelected(item);
+
+            case Window.FEATURE_CONTEXT_MENU:
+                return mFragments.dispatchContextItemSelected(item);
+
+            default:
+                return false;
+        }
     }
 
     @Override
@@ -355,6 +680,8 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
     public void onPanelClosed(int featureId, android.view.Menu menu) {
         switch (featureId) {
             case Window.FEATURE_OPTIONS_PANEL:
+                mFragments.dispatchOptionsMenuClosed(new MenuWrapper(menu));
+
                 if (!IS_HONEYCOMB && (getSupportActionBar() != null)) {
                     if (DEBUG) Log.d(TAG, "onPanelClosed(int, android.view.Menu): Dispatch menu visibility false to custom action bar.");
                     ((ActionBarImpl)mActionBar).onMenuVisibilityChanged(false);
@@ -362,6 +689,42 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
                 break;
         }
         super.onPanelClosed(featureId, menu);
+    }
+
+    /**
+     * Dispatch onPause() to fragments.
+     */
+    @Override
+    protected void onPause() {
+        super.onPause();
+        mResumed = false;
+        if (mHandler.hasMessages(MSG_RESUME_PENDING)) {
+            mHandler.removeMessages(MSG_RESUME_PENDING);
+            mFragments.dispatchResume();
+        }
+        mFragments.dispatchPause();
+    }
+
+    /**
+     * Dispatch onResume() to fragments.
+     */
+    @Override
+    protected void onResume() {
+        super.onResume();
+        mHandler.sendEmptyMessage(MSG_RESUME_PENDING);
+        mResumed = true;
+        mFragments.execPendingActions();
+    }
+
+    /**
+     * Dispatch onResume() to fragments.
+     */
+    @Override
+    protected void onPostResume() {
+        super.onPostResume();
+        mHandler.removeMessages(MSG_RESUME_PENDING);
+        mFragments.dispatchResume();
+        mFragments.execPendingActions();
     }
 
     @Override
@@ -386,6 +749,10 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
                 if (DEBUG) Log.d(TAG, "onPrepareOptionsMenu(android.view.Menu): Calling support method with custom menu.");
                 prepareResult = onPrepareOptionsMenu(mSupportMenu);
                 if (DEBUG) Log.d(TAG, "onPrepareOptionsMenu(android.view.Menu): Support method result returned " + prepareResult);
+                if (prepareResult) {
+                    if (DEBUG) Log.d(TAG, "onPrepareOptionsMenu(android.view.Menu): Dispatching fragment method with custom menu.");
+                    mFragments.dispatchPrepareOptionsMenu(mSupportMenu);
+                }
             }
 
             if (mOptionsMenuInvalidated) {
@@ -416,6 +783,10 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
             if (DEBUG) Log.d(TAG, "onPrepareOptionsMenu(android.view.Menu): Calling support method with wrapped native menu.");
             final MenuWrapper wrappedMenu = new MenuWrapper(menu);
             result = onPrepareOptionsMenu(wrappedMenu);
+            if (result) {
+                if (DEBUG) Log.d(TAG, "onPrepareOptionsMenu(android.view.Menu): Dispatching fragment method with wrapped native menu.");
+                mFragments.dispatchPrepareOptionsMenu(wrappedMenu);
+            }
         }
 
         if (DEBUG) Log.d(TAG, "onPrepareOptionsMenu(android.view.Menu): Returning " + result);
@@ -466,11 +837,105 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
      */
     @Override
     public final Object onRetainNonConfigurationInstance() {
+        if (mStopped) {
+            doReallyStop(true);
+        }
+
         Object custom = onRetainCustomNonConfigurationInstance();
-        if (custom == null) {
+
+        ArrayList<Fragment> fragments = mFragments.retainNonConfig();
+        boolean retainLoaders = false;
+        if (mAllLoaderManagers != null) {
+            // prune out any loader managers that were already stopped and so
+            // have nothing useful to retain.
+            for (int i=mAllLoaderManagers.size()-1; i>=0; i--) {
+                LoaderManagerImpl lm = mAllLoaderManagers.valueAt(i);
+                if (lm.mRetaining) {
+                    retainLoaders = true;
+                } else {
+                    lm.doDestroy();
+                    mAllLoaderManagers.removeAt(i);
+                }
+            }
+        }
+        if (fragments == null && !retainLoaders && custom == null) {
             return null;
         }
-        return custom;
+
+        NonConfigurationInstances nci = new NonConfigurationInstances();
+        nci.activity = null;
+        nci.custom = custom;
+        nci.children = null;
+        nci.fragments = fragments;
+        nci.loaders = mAllLoaderManagers;
+        return nci;
+    }
+
+    /**
+     * Save all appropriate fragment state.
+     */
+    @Override
+    protected void onSaveInstanceState(Bundle outState) {
+        super.onSaveInstanceState(outState);
+        Parcelable p = mFragments.saveAllState();
+        if (p != null) {
+            outState.putParcelable(FRAGMENTS_TAG, p);
+        }
+    }
+
+    /**
+     * Dispatch onStart() to all fragments.  Ensure any created loaders are
+     * now started.
+     */
+    @Override
+    protected void onStart() {
+        super.onStart();
+
+        mStopped = false;
+        mReallyStopped = false;
+        mHandler.removeMessages(MSG_REALLY_STOPPED);
+
+        if (!mCreated) {
+            mCreated = true;
+            ensureSupportActionBarAttached(); //Needed for retained fragments
+            mFragments.dispatchActivityCreated();
+        }
+
+        mFragments.noteStateNotSaved();
+        mFragments.execPendingActions();
+
+        if (!mLoadersStarted) {
+            mLoadersStarted = true;
+            if (mLoaderManager != null) {
+                mLoaderManager.doStart();
+            } else if (!mCheckedForLoaderManager) {
+                mLoaderManager = getLoaderManager(-1, mLoadersStarted, false);
+            }
+            mCheckedForLoaderManager = true;
+        }
+        // NOTE: HC onStart goes here.
+
+        mFragments.dispatchStart();
+        if (mAllLoaderManagers != null) {
+            for (int i=mAllLoaderManagers.size()-1; i>=0; i--) {
+                LoaderManagerImpl lm = mAllLoaderManagers.valueAt(i);
+                lm.finishRetain();
+                lm.doReportStart();
+            }
+        }
+    }
+
+    /**
+     * Dispatch onStop() to all fragments.  Ensure all loaders are stopped.
+     */
+    @Override
+    protected void onStop() {
+        super.onStop();
+
+        mStopped = true;
+        mHandler.sendEmptyMessage(MSG_REALLY_STOPPED);
+
+        mFragments.dispatchStop();
     }
 
     /**
@@ -513,7 +978,9 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
      * {@link #onRetainCustomNonConfigurationInstance()}.
      */
     public Object getLastCustomNonConfigurationInstance() {
-        return getLastNonConfigurationInstance();
+        NonConfigurationInstances nc = (NonConfigurationInstances)
+                getLastNonConfigurationInstance();
+        return nc != null ? nc.custom : null;
     }
 
     /**
@@ -542,6 +1009,53 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
         }
         writer.print(prefix); writer.print("Local FragmentActivity ");
                 writer.print(Integer.toHexString(System.identityHashCode(this)));
+                writer.println(" State:");
+        String innerPrefix = prefix + "  ";
+        writer.print(innerPrefix); writer.print("mCreated=");
+                writer.print(mCreated); writer.print("mResumed=");
+                writer.print(mResumed); writer.print(" mStopped=");
+                writer.print(mStopped); writer.print(" mReallyStopped=");
+                writer.println(mReallyStopped);
+        writer.print(innerPrefix); writer.print("mLoadersStarted=");
+                writer.println(mLoadersStarted);
+        if (mLoaderManager != null) {
+            writer.print(prefix); writer.print("Loader Manager ");
+                    writer.print(Integer.toHexString(System.identityHashCode(mLoaderManager)));
+                    writer.println(":");
+            mLoaderManager.dump(prefix + "  ", fd, writer, args);
+        }
+        mFragments.dump(prefix, fd, writer, args);
+    }
+
+    void doReallyStop(boolean retaining) {
+        if (!mReallyStopped) {
+            mReallyStopped = true;
+            mRetaining = retaining;
+            mHandler.removeMessages(MSG_REALLY_STOPPED);
+            onReallyStop();
+        }
+    }
+
+    /**
+     * Pre-HC, we didn't have a way to determine whether an activity was
+     * being stopped for a config change or not until we saw
+     * onRetainNonConfigurationInstance() called after onStop().  However
+     * we need to know this, to know whether to retain fragments.  This will
+     * tell us what we need to know.
+     */
+    void onReallyStop() {
+        if (mLoadersStarted) {
+            mLoadersStarted = false;
+            if (mLoaderManager != null) {
+                if (!mRetaining) {
+                    mLoaderManager.doStop();
+                } else {
+                    mLoaderManager.doRetain();
+                }
+            }
+        }
+
+        mFragments.dispatchReallyStop();
     }
 
     // ------------------------------------------------------------------------
@@ -642,7 +1156,19 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
      */
     @Override
     public FragmentManager getSupportFragmentManager() {
-        return null;
+        return mFragments;
+    }
+
+    /**
+     * Modifies the standard behavior to allow results to be delivered to fragments.
+     * This imposes a restriction that requestCode be <= 0xffff.
+     */
+    @Override
+    public void startActivityForResult(Intent intent, int requestCode) {
+        if (requestCode != -1 && (requestCode&0xffff0000) != 0) {
+            throw new IllegalArgumentException("Can only use lower 16 bits for requestCode");
+        }
+        super.startActivityForResult(intent, requestCode);
     }
 
     /**
@@ -663,6 +1189,13 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
 
     void invalidateSupportFragmentIndex(int index) {
         //Log.v(TAG, "invalidateFragmentIndex: index=" + index);
+        if (mAllLoaderManagers != null) {
+            LoaderManagerImpl lm = mAllLoaderManagers.get(index);
+            if (lm != null && !lm.mRetaining) {
+                lm.doDestroy();
+                mAllLoaderManagers.remove(index);
+            }
+        }
     }
 
     // ------------------------------------------------------------------------
@@ -674,6 +1207,27 @@ public class SherlockPreferenceActivity extends PreferenceActivity implements Su
      */
     @Override
     public LoaderManager getSupportLoaderManager() {
-        return null;
+        if (mLoaderManager != null) {
+            return mLoaderManager;
+        }
+        mCheckedForLoaderManager = true;
+        mLoaderManager = getLoaderManager(-1, mLoadersStarted, true);
+        return mLoaderManager;
+    }
+
+    LoaderManagerImpl getLoaderManager(int index, boolean started, boolean create) {
+        if (mAllLoaderManagers == null) {
+            mAllLoaderManagers = new HCSparseArray<LoaderManagerImpl>();
+        }
+        LoaderManagerImpl lm = mAllLoaderManagers.get(index);
+        if (lm == null) {
+            if (create) {
+                lm = new LoaderManagerImpl(this, started);
+                mAllLoaderManagers.put(index, lm);
+            }
+        } else {
+            lm.updateActivity(this);
+        }
+        return lm;
     }
 }